backend/requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
flask==3.0.0
Flask-Cors==4.0.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
localStoragePy==0.2.3
mariadb==1.1.8
MarkupSafe==2.1.3
packaging==23.2
PyYAML==6.0.1
requests==2.31.0
urllib3==2.1.0
werkzeug==3.0.1
zipp==3.17.0
